+++ server/src/main/kotlin/suwayomi/tachidesk/graphql/types/DownloadType.kt
@@ -35,7 +35,7 @@ data class DownloadStatus(
             Status.Stopped -> DownloaderState.STOPPED
             Status.Started -> DownloaderState.STARTED
         },
-        downloadStatus.queue.map { DownloadType(it) },
+        downloadStatus.queue.mapIndexed { index, item -> DownloadType(item, index) },
     )
 }
 
@@ -56,7 +56,7 @@ data class DownloadUpdates(
             Status.Started -> DownloaderState.STARTED
         },
         downloadUpdates.updates.map { DownloadUpdate(it) },
-        downloadUpdates.initial?.map { DownloadType(it) },
+        downloadUpdates.initial?.mapIndexed { index, item -> DownloadType(item, index) },
         omittedUpdates,
     )
 }
@@ -71,7 +71,7 @@ class DownloadType(
     val tries: Int,
     val position: Int,
 ) : Node {
-    constructor(downloadChapter: DownloadQueueItem) : this(
+    constructor(downloadChapter: DownloadQueueItem, position: Int) : this(
         downloadChapter.chapterId,
         downloadChapter.mangaId,
         when (downloadChapter.state) {
@@ -82,7 +82,7 @@ class DownloadType(
         },
         downloadChapter.progress,
         downloadChapter.tries,
-        downloadChapter.position,
+        position,
     )
 
     fun manga(dataFetchingEnvironment: DataFetchingEnvironment): CompletableFuture<MangaType> {
@@ -110,7 +110,7 @@ class DownloadUpdate(
 ) : Node {
     constructor(downloadUpdate: DownloadUpdate) : this(
         downloadUpdate.type,
-        DownloadType(downloadUpdate.downloadQueueItem),
+        DownloadType(downloadUpdate.downloadQueueItem, downloadUpdate.position),
     )
 }