+++ server/src/main/kotlin/suwayomi/tachidesk/manga/impl/download/model/DownloadQueueItem.kt
@@ -14,7 +14,6 @@ class DownloadQueueItem(
     val chapterIndex: Int,
     val mangaId: Int,
     val sourceId: Long,
-    var position: Int,
     var pageCount: Int,
     var state: DownloadState = Queued,
     var progress: Float = 0f,